+++ x-pack/platform/packages/shared/dashboard-agent/dashboard-agent-common/dashboard_state_helpers.ts
@@ -1,0 +1,26 @@
+export const DEFAULT_TIME_RANGE = { from: 'now-24h', to: 'now' } as const;
+
+/**
+ * Default values for all dashboard state fields except project_routing.
+ */
+export const EMPTY_DASHBOARD_STATE = Object.freeze({
+  title: '',
+  description: '',
+  panels: [],
+  time_range: DEFAULT_TIME_RANGE,
+  query: { expression: '', language: 'kql' as const },
+  filters: [],
+  options: {
+    hide_panel_titles: false,
+    hide_panel_borders: false,
+    use_margins: true,
+    auto_apply_filters: true,
+    sync_colors: false,
+    sync_cursor: true,
+    sync_tooltips: false,
+  },
+  pinned_panels: [],
+  refresh_interval: { pause: true, value: 0 },
+  tags: [],
+  access_control: {},
+});
